requirements.txt
chromadb
open-clip-torch
transformers
python-dotenv
flask
